# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghayth ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard stats
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: classification
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "200":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClientResponse"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # ─── HR ────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        assignments:
          type: array
          items:
            type: object
      required: [token, assignments]
...
    DashboardData:
      type: object
      properties:
        cards:
          type: object
          properties:
            todayTasks:
              type: integer
            awaitingMe:
              type: integer
            overdue:
              type: integer
            completedPct:
              type: integer
        todayTasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        pendingApprovals:
          type: array
          items:
            type: object
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
      required: [cards, todayTasks, pendingApprovals, notifications]
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        classification:
          type: string
        source:
          type: string
        totalRevenue:
          type: number
        isBlacklisted:
          type: boolean
        assignedToName:
          type: string
        createdAt:
          type: string
      required: [id, classification]
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        source:
          type: string
      required: [phone]
    CreateClientResponse:
      type: object
      properties:
        clientId:
          type: integer
        code:
          type: string
      required: [clientId, code]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: string
        priority:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, type, title, isRead]
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        status:
          type: string
        priority:
          type: string
        scheduledStart:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        createdAt:
          type: string
      required: [id, title, type, status, priority]
...